+++ packages/pdf-viewer/demo/DemoDev.tsx
@@ -5,7 +5,7 @@
  */
 
 import React from 'react';
-import PdfViewer, { PDFOpenFileButtonPlugin, PDFPainterPlugin, PDFTooltipPlugin, usePdfViewerRef } from '@orca-fe/pdf-viewer';
+import PdfViewer, { PDFContextPrintPlugin, PDFOpenFileButtonPlugin, PDFPainterPlugin, PDFTooltipPlugin, usePdfViewerRef } from '@orca-fe/pdf-viewer';
 
 const Demo1 = () => {
   const pdfViewerRef = usePdfViewerRef();
@@ -28,6 +28,7 @@ const Demo1 = () => {
         <PDFOpenFileButtonPlugin />
         <PDFPainterPlugin />
         <PDFTooltipPlugin />
+        <PDFContextPrintPlugin />
       </PdfViewer>
     </div>
   );