+++ packages/pdf-viewer/src/plugins/PDFContextMenuPlugin/index.ts
@@ -1,0 +1,5 @@
+import PDFContextMenuPlugin from './PDFContextMenuPlugin';
+
+export default PDFContextMenuPlugin;
+
+export * from './PDFContextMenuPlugin';