+++ packages/pdf-viewer/src/plugins/PDFContextPrintPlugin/index.ts
@@ -1,0 +1,5 @@
+import PdfContextPrintPlugin from './PdfContextPrintPlugin';
+
+export default PdfContextPrintPlugin;
+
+export * from './PdfContextPrintPlugin';